+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Replace sprintf with text templates. (2022-11-09, @bamdadnouri, !57)
 * Add query builder layer. (2022-11-13, @n25a, !58)
+* Add function builder layer. (2022-11-17, @n25a, !59)
 
 # v2.0.0 - Nov 08 2022